Update: Average star rating

+++ src/pages/ProductsPage.tsx
@@ -103,7 +103,7 @@ const ProductsPage = () => {
                 <div>
                     <span className='text-xs'>+{vals.productDetails?.length || 0} kích thước</span>
                     <p className='text-md my-1 line-clamp-2'>{vals.nameProduct}</p>
-                    <Rate className='my-1' disabled value={Number.parseInt(vals.averageRating)} />
+                    <Rate className='my-1' disabled value={Number.parseFloat(vals.averageRating)} />
                     <div className='flex items-center gap-2'>
                         <span className='text-red-500 font-semibold text-sm'>
                             {formatCurrency(vals.productDetails[0].promotionalPrice)}
